+++ tkinter-app/coco.names
@@ -1,0 +1,80 @@
+persona
+bicicleta
+coche
+moto
+avion
+autobus
+tren
+camion
+barco
+semaforo
+boca de incendio
+señal de stop
+parquimetro
+banco
+pajaro
+gato
+perro
+caballo
+oveja
+vaca
+elefante
+oso
+cebra
+jirafa
+mochila
+paraguas
+bolso
+Corbata
+maleta
+frisbee
+esquis
+snowboard
+pelota deportiva
+cometa
+bate de beisbol
+guante de béisbol
+patineta
+tabla de surf
+raqueta de tenis
+botella
+Copa de vino
+vaso
+tenedor
+cuchillo
+cuchara
+cuenco
+platano
+manzana
+emparedado
+naranja
+brocoli
+Zanahoria
+Pancho
+Pizza
+rosquilla
+pastel
+silla
+sofa
+planta en maceta
+cama
+comedor
+baño
+tvmonitor
+ordenador portatil
+raton
+remoto
+teclado
+Telefono
+microonda
+horno
+tostadora
+lavabo
+refrigerador
+libro
+reloj
+florero
+tijeras
+oso de peluche
+Secadora de pelo
+cepillo de dientes
